refactor: Adopt new storybook config format
