+++ .storybook/theme.js
@@ -1,0 +1,10 @@
+import {themes, create} from '@storybook/theming';
+import {commonColors, typeColors, fontFamily} from '../modules/core/react';
+
+export default create({
+  ...themes.light, // Overrides a user's preferred color scheme (e.g. Dark Mode), will need to flesh this out later when CK is compatible with a dark mode
+  brandTitle: 'Canvas Kit',
+  mainTextColor: typeColors.body,
+  mainTextFace: fontFamily,
+  mainBackground: commonColors.backgroundAlt,
+});
