+++ package.json
@@ -14,14 +14,16 @@
     "@commitlint/cli": "^8.3.5",
     "@commitlint/config-conventional": "^8.1.0",
     "@emotion/is-prop-valid": "^0.8.2",
-    "@storybook/addon-actions": "^5.3.19",
-    "@storybook/addon-docs": "^5.3.19",
-    "@storybook/addon-knobs": "^5.3.19",
-    "@storybook/addon-storysource": "^5.3.19",
-    "@storybook/addon-viewport": "^5.3.19",
-    "@storybook/components": "^5.3.19",
-    "@storybook/react": "^5.3.19",
-    "@storybook/source-loader": "^5.3.19",
+    "@storybook/addon-actions": "^6.0.27",
+    "@storybook/addon-docs": "^6.0.27",
+    "@storybook/addon-essentials": "^6.0.28",
+    "@storybook/addon-knobs": "^6.0.27",
+    "@storybook/addon-storysource": "^6.0.27",
+    "@storybook/addon-viewport": "^6.0.27",
+    "@storybook/components": "^6.0.27",
+    "@storybook/preset-scss": "^1.0.3",
+    "@storybook/react": "^6.0.27",
+    "@storybook/source-loader": "^6.0.27",
     "@testing-library/cypress": "^6.0.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^9.1.4",
